B2B-1610: Upgrade to flysystem 2.0
- test fix

+++ dev/tests/integration/testsuite/Magento/AwsS3/Driver/AwsS3PerfTest.php
@@ -146,6 +146,8 @@ class AwsS3PerfTest extends \PHPUnit\Framework\TestCase
 
     public static function tearDownAfterClass(): void
     {
-        self::$awsDriver->deleteDirectory(self::TEST_FILE_DIR);
+        if (self::$awsDriver) {
+            self::$awsDriver->deleteDirectory(self::TEST_FILE_DIR);
+        }
     }
 }
